Add support for /etc/pkgtools/reject.conf
The format of the file is a single regex per line matching
the entries in a package.  The entries in a package start with
'./' which is skipped for the regex matching so in essence a
regex ^etc/.*$ will match etc/passwd which appears in the package
as ./etc/passwd.

Keep in mind that regex matches are only checked against files
in a package and *not* the filesystem.

All commands obey the reject matches.

installpkg and installpkg -f will *not* install a file if there is
a regex match for that file.

removepkg and removepkg -f will *not* remove a file if there is a
regex match for that file.

Filesystem collisions for rejected files will still be reported
installpkg -f will bypass those checks and it will still reject the
files.

An upcoming patch will move the rejected files to some other place
on the filesystem.

The database entry for the package will include all the original
files even if some of them were rejected.  It is up to the user to ensure
that rejected files are properly handled.

Notice that installpkg will reject files even if they do not exist
on the filesystem.  The rejection operation applies when installing,
removing and upgrading packages.